CIV-15230 add abstract classes
