+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/camunda/notification/AbstractNotifyManageStayHandler.java
@@ -3,6 +3,7 @@ package uk.gov.hmcts.reform.civil.handler.callback.camunda.notification;
 import lombok.RequiredArgsConstructor;
 import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackResponse;
+import uk.gov.hmcts.reform.civil.callback.Callback;
 import uk.gov.hmcts.reform.civil.callback.CallbackHandler;
 import uk.gov.hmcts.reform.civil.callback.CallbackParams;
 import uk.gov.hmcts.reform.civil.model.CaseData;
@@ -11,11 +12,18 @@ import uk.gov.hmcts.reform.civil.notify.NotificationsProperties;
 
 import java.util.Map;
 
+import static uk.gov.hmcts.reform.civil.callback.CallbackType.ABOUT_TO_SUBMIT;
+
 @RequiredArgsConstructor
 public abstract class AbstractNotifyManageStayHandler extends CallbackHandler implements NotificationData {
 
     protected final NotificationService notificationService;
     protected final NotificationsProperties notificationsProperties;
+
+    @Override
+    public Map<String, Callback> callbacks() {
+        return Map.of(callbackKey(ABOUT_TO_SUBMIT), this::sendNotification);
+    }
 
     public CallbackResponse sendNotification(CallbackParams callbackParams) {
         CaseData caseData = callbackParams.getCaseData();
