+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/camunda/notification/NotifyClaimantStayLiftedHandler.java
@@ -1,21 +1,17 @@
 package uk.gov.hmcts.reform.civil.handler.callback.camunda.notification;
 
 import org.springframework.stereotype.Service;
-import uk.gov.hmcts.reform.civil.callback.Callback;
 import uk.gov.hmcts.reform.civil.callback.CallbackParams;
 import uk.gov.hmcts.reform.civil.callback.CaseEvent;
-import uk.gov.hmcts.reform.civil.model.CaseData;
 import uk.gov.hmcts.reform.civil.notify.NotificationService;
 import uk.gov.hmcts.reform.civil.notify.NotificationsProperties;
 
 import java.util.List;
-import java.util.Map;
 
-import static uk.gov.hmcts.reform.civil.callback.CallbackType.ABOUT_TO_SUBMIT;
 import static uk.gov.hmcts.reform.civil.callback.CaseEvent.NOTIFY_CLAIMANT_STAY_LIFTED;
 
 @Service
-public class NotifyClaimantStayLiftedHandler extends AbstractNotifyManageStayHandler {
+public class NotifyClaimantStayLiftedHandler extends AbstractNotifyManageStayClaimantHandler {
 
     private static final String TASK_ID = "NotifyClaimantStayLifted";
     private static final String REFERENCE_TEMPLATE = "stay-lifted-claimant-notification-%s";
@@ -36,31 +32,8 @@ public class NotifyClaimantStayLiftedHandler extends AbstractNotifyManageStayHan
     }
 
     @Override
-    protected String getRecipient(CallbackParams callbackParams) {
-        CaseData caseData = callbackParams.getCaseData();
-        return caseData.isApplicantLiP()
-            ? caseData.getClaimantUserDetails().getEmail()
-            : caseData.getApplicantSolicitor1UserDetails().getEmail();
-    }
-
-    @Override
-    protected boolean isLiP(CaseData caseData) {
-        return caseData.isApplicantLiP();
-    }
-
-    @Override
     public List<CaseEvent> handledEvents() {
         return EVENTS;
     }
 
-    @Override
-    protected String getPartyName(CallbackParams callbackParams) {
-        CaseData caseData = callbackParams.getCaseData();
-        return caseData.getApplicant1().getPartyName();
-    }
-
-    @Override
-    public Map<String, Callback> callbacks() {
-        return Map.of(callbackKey(ABOUT_TO_SUBMIT), this::sendNotification);
-    }
 }
