+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/camunda/notification/NotifyClaimantStayUpdateRequestedHandler.java
@@ -1,7 +1,6 @@
 package uk.gov.hmcts.reform.civil.handler.callback.camunda.notification;
 
 import org.springframework.stereotype.Service;
-import uk.gov.hmcts.reform.civil.callback.Callback;
 import uk.gov.hmcts.reform.civil.callback.CallbackParams;
 import uk.gov.hmcts.reform.civil.callback.CaseEvent;
 import uk.gov.hmcts.reform.civil.model.CaseData;
@@ -9,13 +8,11 @@ import uk.gov.hmcts.reform.civil.notify.NotificationService;
 import uk.gov.hmcts.reform.civil.notify.NotificationsProperties;
 
 import java.util.List;
-import java.util.Map;
 
-import static uk.gov.hmcts.reform.civil.callback.CallbackType.ABOUT_TO_SUBMIT;
 import static uk.gov.hmcts.reform.civil.callback.CaseEvent.NOTIFY_CLAIMANT_STAY_UPDATE_REQUESTED;
 
 @Service
-public class NotifyClaimantStayUpdateRequestedHandler extends AbstractNotifyManageStayHandler {
+public class NotifyClaimantStayUpdateRequestedHandler extends AbstractNotifyManageStayClaimantHandler {
 
     private static final String TASK_ID = "NotifyClaimantStayUpdateRequested";
     private static final String REFERENCE_TEMPLATE = "stay-update-requested-claimant-notification-%s";
@@ -23,11 +20,6 @@ public class NotifyClaimantStayUpdateRequestedHandler extends AbstractNotifyMana
 
     public NotifyClaimantStayUpdateRequestedHandler(NotificationService notificationService, NotificationsProperties notificationsProperties) {
         super(notificationService, notificationsProperties);
-    }
-
-    @Override
-    public Map<String, Callback> callbacks() {
-        return Map.of(callbackKey(ABOUT_TO_SUBMIT), this::sendNotification);
     }
 
     @Override
@@ -41,14 +33,6 @@ public class NotifyClaimantStayUpdateRequestedHandler extends AbstractNotifyMana
     }
 
     @Override
-    protected String getRecipient(CallbackParams callbackParams) {
-        CaseData caseData = callbackParams.getCaseData();
-        return caseData.isApplicantLiP()
-            ? caseData.getClaimantUserDetails().getEmail()
-            : caseData.getApplicantSolicitor1UserDetails().getEmail();
-    }
-
-    @Override
     protected String getNotificationTemplate(CaseData caseData) {
         if (isLiP(caseData)) {
             // TODO: add lip template
@@ -59,18 +43,8 @@ public class NotifyClaimantStayUpdateRequestedHandler extends AbstractNotifyMana
     }
 
     @Override
-    protected boolean isLiP(CaseData caseData) {
-        return caseData.isApplicantLiP();
-    }
-
-    @Override
     public List<CaseEvent> handledEvents() {
         return EVENTS;
     }
 
-    @Override
-    protected String getPartyName(CallbackParams callbackParams) {
-        CaseData caseData = callbackParams.getCaseData();
-        return caseData.getApplicant1().getPartyName();
-    }
 }
