+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/camunda/notification/NotifyDefendantStayUpdateRequestedHandler.java
@@ -1,7 +1,6 @@
 package uk.gov.hmcts.reform.civil.handler.callback.camunda.notification;
 
 import org.springframework.stereotype.Service;
-import uk.gov.hmcts.reform.civil.callback.Callback;
 import uk.gov.hmcts.reform.civil.callback.CallbackParams;
 import uk.gov.hmcts.reform.civil.callback.CaseEvent;
 import uk.gov.hmcts.reform.civil.model.CaseData;
@@ -9,15 +8,12 @@ import uk.gov.hmcts.reform.civil.notify.NotificationService;
 import uk.gov.hmcts.reform.civil.notify.NotificationsProperties;
 
 import java.util.List;
-import java.util.Map;
 
-import static java.util.Objects.nonNull;
-import static uk.gov.hmcts.reform.civil.callback.CallbackType.ABOUT_TO_SUBMIT;
 import static uk.gov.hmcts.reform.civil.callback.CaseEvent.NOTIFY_DEFENDANT2_STAY_UPDATE_REQUESTED;
 import static uk.gov.hmcts.reform.civil.callback.CaseEvent.NOTIFY_DEFENDANT_STAY_UPDATE_REQUESTED;
 
 @Service
-public class NotifyDefendantStayUpdateRequestedHandler extends AbstractNotifyManageStayHandler {
+public class NotifyDefendantStayUpdateRequestedHandler extends AbstractNotifyManageStayDefendantHandler {
 
     private static final String TASK_ID = "NotifyDefendantStayUpdateRequested";
     private static final String TASK_ID_DEFENDANT_2 = "NotifyDefendant2StayUpdateRequested";
@@ -35,24 +31,6 @@ public class NotifyDefendantStayUpdateRequestedHandler extends AbstractNotifyMan
     }
 
     @Override
-    protected String getRecipient(CallbackParams callbackParams) {
-        CaseData caseData = callbackParams.getCaseData();
-        if (caseData.isRespondent1LiP() && nonNull(caseData.getRespondent1().getPartyEmail())) {
-            return caseData.getRespondent1().getPartyEmail();
-        } else {
-            if (isRespondentSolicitor2(callbackParams)) {
-                return caseData.getRespondentSolicitor2EmailAddress();
-            }
-            return caseData.getRespondentSolicitor1EmailAddress();
-        }
-    }
-
-    @Override
-    protected boolean isLiP(CaseData caseData) {
-        return caseData.isRespondent1LiP();
-    }
-
-    @Override
     protected String getNotificationTemplate(CaseData caseData) {
         if (isLiP(caseData)) {
             // TODO: add lip template
@@ -63,16 +41,8 @@ public class NotifyDefendantStayUpdateRequestedHandler extends AbstractNotifyMan
     }
 
     @Override
-    protected String getPartyName(CallbackParams callbackParams) {
-        CaseData caseData = callbackParams.getCaseData();
-        return isRespondentSolicitor2(callbackParams)
-            ? caseData.getRespondent2().getPartyName()
-            : caseData.getRespondent1().getPartyName();
-    }
-
-    @Override
-    public Map<String, Callback> callbacks() {
-        return Map.of(callbackKey(ABOUT_TO_SUBMIT), this::sendNotification);
+    public List<CaseEvent> handledEvents() {
+        return EVENTS;
     }
 
     @Override
@@ -85,11 +55,7 @@ public class NotifyDefendantStayUpdateRequestedHandler extends AbstractNotifyMan
     }
 
     @Override
-    public List<CaseEvent> handledEvents() {
-        return EVENTS;
-    }
-
-    private boolean isRespondentSolicitor2(CallbackParams callbackParams) {
+    protected boolean isRespondentSolicitor2(CallbackParams callbackParams) {
         CaseEvent caseEvent = CaseEvent.valueOf(callbackParams.getRequest().getEventId());
         return NOTIFY_DEFENDANT2_STAY_UPDATE_REQUESTED.equals(caseEvent);
     }
